server/ml-backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
numpy==2.3.1
Pillow==10.4.0
tensorflow==2.15.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0 